implemented rbac on citl revision

+++ pages/api/citl_revision/[id]/index.ts
@@ -1,9 +1,6 @@
 import prisma from "@/prisma/client";
-import cITLRevisionAbility from "@/services/ability/cITLRevisionAbility";
 import getServerUser from "@/services/getServerUser";
 import logger from "@/services/logger";
-import { ForbiddenError } from "@casl/ability";
-import { accessibleBy } from "@casl/prisma";
 import { User } from "@prisma/client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import * as Yup from "yup";
@@ -20,7 +17,6 @@ export default async function handler(
     logger.error(error);
     return res.status(401).json({ error: { message: "Unauthorized" } });
   }
-  const ability = cITLRevisionAbility({ user });
 
   const getHandler = async () => {
     try {
@@ -31,19 +27,17 @@ export default async function handler(
       await validator.validate(req.query);
 
       const { id } = validator.cast(req.query);
-      const cITLRevision =
-        await prisma.cITLRevision.findFirstOrThrow({
-          where: {
-            AND: [
-              accessibleBy(ability).CITLRevision,
-              {
-                id: {
-                  equals: id,
-                },
+      const cITLRevision = await prisma.cITLRevision.findFirstOrThrow({
+        where: {
+          AND: [
+            {
+              id: {
+                equals: id,
               },
-            ],
-          },
-        });
+            },
+          ],
+        },
+      });
 
       return res.json(cITLRevision);
     } catch (error: any) {
@@ -62,12 +56,15 @@ export default async function handler(
 
       await validator.validate(req.query);
 
-      ForbiddenError.from(ability).throwUnlessCan(
-        "delete",
-        "CITLRevision"
-      );
-
       const { id } = validator.cast(req.query);
+
+      if (!user.isAdmin) {
+        return res.status(403).json({
+          error: {
+            message: "You are not allowed to perform this action",
+          },
+        });
+      }
 
       const cITLRevision = await prisma.cITLRevision.delete({
         where: {
@@ -83,7 +80,6 @@ export default async function handler(
         .json({ error: { message: error?.message ?? "Server Error" } });
     }
   };
-
 
   switch (req.method) {
     case "DELETE":
